First version: working root/user split, comms protocol, and UI framework
Set up the application structure, including handling running as root
and dropping privs. Also includes a helper process which can be run as
root (to get privileged access to the filesystems) and can be run
remotely (for systems with no GUI access). The two parts of the system
communicate with a simple text-based protocol over stdin/stdout.

Implements a scan for filesystems as a simple test of the user/root
separation. No GUI interaction yet.