Safer wrapping for open filesystems
Opening a filesystem requires closing it again afterwards, regardless
of the exit method taken from the method that opened it. Python
provides context managers to do precisely this sort of thing, so
create a context manager to wrap up opening, holding and closing a
filehandle on a filesystem safely regardless of whether there are
exceptions thrown on the way.

If we don't do this, if we receive an exception while playing with an
open file handle, the filehandle remains open, and we can't then
unmount the filesystem cleanly when we exit.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>