Process array data in a pipelined way
In the GUI<->helper protocol, returning array data is a way of returning
potentially large data without having it fully held in memory on either
side of the connection. To support this streaming mode, we need something
that can read and process the data piecemeal, so implement
Requester.request_array() using a generator function. This means that we
can treat the returned data as an iterator over objects.

The change to a generator means that we don't actually know the outcome
code until the end of the sequence, so we immediately return with an
undecided return code, and raise an exception if we hit something that we
can't deal with.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>