helper: Implement "ls" command for inspecting filesystem objects
We need to be able to list the contents of directories on the remote
filesystem, in order to give the user a navigation over the filesystem
to select where to create subvolumes. We implement helper-side filters,
so that for large directories we can reduce the amount of data sent
across.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>