Improve communications protocol parsing
As it stands, the parts of the command sent from the GUI to the root helper
are separated by whitespace. This will cause problems in the parsing of
commands involving more than one pathname, or with a pathname that is not
at the end of the line, because we won't know where the bounds of the
pathname lie.

Refactor the parser for the helper to parse every command line as separated
by spaces only (i.e. tabs and newlines are not separators) *and* to quote
characters with a backslash (so that spaces can be quoted without being
separators.

Also refactor the GUI-side functions so that we pass individual elements of
the command line in, allowing for them to be properly quoted and separated
before being sent to the helper.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>