Move common code out to a decorator function
tk doesn't allow us to pass parameters to menu-item handlers, so each
handler needs to use the same piece of code to get the current selection
in the list that we care about. We can factor out that piece of code as
a decorator, and wrap up our functions so that they look like they take
parameters of the currently selected subvolume.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>