Deal with a potential division by zero
In some broken filesystems, the size of some or all of the top-level
size boxes may be zero. In this case, there will be a division-by-zero
error when the boxes are scaled to a suitable size. If the total size
of a box is found to be zero, set it to 1 instead to avoid this
problem. (This should have no effect on the display of the results,
since in this case the rectangles in question will be of size zero
anyway).

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>