Warn once on missing icons
We already warn on stdout when there are missing icons, but we should
put it in the user's face once, and tell them how to fix it. This
should only show up on freshly-pulled git repos (and occasions when we
add new icons to git). Distribution tarballs will have the pre-built
icons in them.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>