Convert python2 idiom to python3 idiom before installing under python3
We have a mixed-version source, unfortunately: Everything in
btrfsgui/hlp is python2, and everything else is python3. Since
distutils modifies the shebang line at the top of the file, we need to
ensure that we're compatible. Fortunately, few changes need to be
made, and 2to3 gets it right, so we can do the conversion completely
automatically.

We run 2to3 on the relevant files, capture the patch it outputs, run
that through patch with backups set on, do the installation, and then
revert the patch by restoring from the earlier backups. Ugly, but
expedient.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>