Handle selection/deselection of devices in the device list for mkfs
Implement the ability to double-click on a device in the device
list. This selectes it (prefixed with *), and moves it to the top of
the list. The user can de-select it again by double-clicking. Items
are identified internally by their device major/minor number, and the
same major/minor cannot be selected more than once.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>