Add search path for root helper
We were simply looking in the current directory for the helper, which
is clearly not going to work in general. Unlike the half-baked solution
in c8d1e893, this patch works by running a shell in the target environment
(sudo, remote, etc), and modifying the shell's PATH variable with some
likely locations.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>