helper: process-private temp directories
When we spawn off subprocesses for long-lived operations like rm dev
or balance, those processes need their own FS mount to work with, not
the one the helper's trying to use. Therefore, we add a per-pid
subdirectory, so that each new subprocess has its own place to store
data. All subprocesses associated with the same helper appear in the
same temp dir.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>