Fix bug when filesystem scan results change
If the filesystem has changed its fundamental parameters (e.g. the
device list) since the last time we scanned it, these changes are not
reflected in the data we cache for the selected item. This can then
lead to things like the device usage display breaking, or not showing
enough devices. This patch ensures that the data cache of the selected
item is up to date.

Signed-off-by: Hugo Mills <hugo@carfax.org.uk>